+++ db/upgrade.php
@@ -143,6 +143,11 @@ function xmldb_format_grid_upgrade($oldversion = 0) {
         $value = get_config('format_grid', 'defaultsection0ownpagenogridonesection');
         set_config('defaultsetsection0ownpagenogridonesection', $value, 'format_grid');
 
+        global $CFG;
+        require_once($CFG->dirroot.'/course/format/grid/lib.php'); // For format_grid.
+
+        format_grid::update_displayed_images_callback();
+
         upgrade_plugin_savepoint(true, 2020111402, 'format', 'grid');
     }