+++ lang/en/format_grid.php
@@ -159,6 +159,11 @@ $string['setcurrentselectedimagecontainercolour_help'] = 'Set the current select
 $string['setnewactivity'] = 'Show new activity notification image';
 $string['setnewactivity_help'] = "Show the new activity notification image when a new activity or resource are added to a section.";
 
+$string['singlepagesummaryimage'] = 'Show the grid image in the section summary on a single page.';
+$string['singlepagesummaryimage_help'] = "When the 'Course layout' is set to 'Show one section per page' then show the grid image for that section in the section summary when there is a summary in the position stated.";
+$string['defaultsinglepagesummaryimage'] = 'Show the grid image in the section summary on a single page.';
+$string['defaultsinglepagesummaryimage_desc'] = "When the 'Course layout' is set to 'Show one section per page' then show the grid image for that section in the section summary when there is a summary in the position stated.";
+
 $string['setfitsectioncontainertowindow'] = 'Fit the section popup to the window';
 $string['setfitsectioncontainertowindow_help'] = 'If enabled, the popup box with the contents of the section will fit to the size of the window and will scroll inside if necessary.  If disabled, the entire page will scroll instead.';
 
@@ -271,6 +276,10 @@ $string['resetnewactivity'] = 'New activity';
 $string['resetnewactivity_help'] = 'Resets the new activity notification image to follow the site default value.';
 $string['resetallnewactivity'] = 'New activities';
 $string['resetallnewactivity_help'] = 'Resets the new activity notification images to follow the site default value.';
+$string['resetsinglepagesummaryimage'] = 'Single page summary image';
+$string['resetsinglepagesummaryimage_help'] = 'Resets the single page summary image to follow the site default value.';
+$string['resetallsinglepagesummaryimage'] = 'Single page summary images';
+$string['resetallsinglepagesummaryimage_help'] = 'Resets the single page summary images to follow the site default value.';
 $string['resetfitpopup'] = 'Fit section popup to the window';
 $string['resetfitpopup_help'] = 'Resets the \'Fit section popup to the window\' to follow the site default value.';
 $string['resetallfitpopup'] = 'Fit section popups to the window';